test: relocate practice recommendation check
